Consolidate test helpers into a single file
